requirements.txt
fastmcp
requests
httpx
tenacity
pyyaml
gitpython
cryptography
aiohttp
aiohttp-cors
prometheus-client
structlog
python-json-logger
asyncio-throttle
psutil

# Development dependencies
pytest
pytest-asyncio
pytest-cov
pytest-mock 